Caches and clears blob URLs for garbage collection.
